+++ azext_iot/iothub/params.py
@@ -62,7 +62,17 @@ def load_iothub_arguments(self, _):
             options_list=["--auth-type"],
             arg_type=hub_auth_type_dataplane_param_type,
         )
-        context.argument("data", options_list=["--data", "--da"], help="Message body.")
+        context.argument(
+            "data",
+            options_list=["--data", "--da"],
+            help="Message body. Provide text or raw json.")
+        context.argument(
+            "data_file_path",
+            is_preview=True,
+            options_list=["--data-file-path", "--dfp"],
+            help="""Provide path to file for message body payload. Please note when the payload needs
+            to be sent in binary format, set the content type to application/octet-stream."""
+        )
         context.argument(
             "properties",
             options_list=["--properties", "--props", "-p"],